flood_system/requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
flask>=2.0.0
flask-socketio>=5.0.0
pyyaml>=5.4.0
pyserial>=3.5
requests>=2.25.0
gevent>=21.0.0
gevent-websocket>=0.10.0
